+++ CHANGELOG.md
@@ -1,7 +1,13 @@
 ## Changelog
 
-### 0.2
+#### 0.3
+1. Expand tests.
+2. Fix bug where static file not defined in a layer could not be overwritten in a layer.
+3. Provide a layer aware replacement for collectstatic.
+
+#### 0.2
 1. Inevitable package rename.
 
-### 0.1
+#### 0.1
 1. Initial release.
+